add identifier for secret key
use secret key pair to create identifier in such a way
that only owner of key is able to compute the identifier.
Thus it doesn't help in finding correct key when decrypting data,
but it does help users to debug issues, as check -v shows
which key is associated with current user.